feat: improve typescript support for config file
