+++ src/node/index.ts
@@ -4,4 +4,4 @@ export * from './serve/serve'
 export * from './config'
 export * from './markdown/markdown'
 
-export type { SiteData, HeadConfig, LocaleConfig } from '../../types/shared'
+export type { SiteData, HeadConfig, LocaleConfig, DefaultTheme } from '../../types/shared'
