+++ src/shared/shared.ts
@@ -5,7 +5,8 @@ export type {
   PageData,
   HeadConfig,
   LocaleConfig,
-  Header
+  Header,
+  DefaultTheme,
 } from '../../types/shared'
 
 export const EXTERNAL_URL_RE = /^https?:/i
