+++ src/shared/tsconfig.json
@@ -3,5 +3,5 @@
   "compilerOptions": {
     "baseUrl": "."
   },
-  "include": [".", "../../types/shared.d.ts"]
+  "include": ["."]
 }
