+++ types/index.d.ts
@@ -1,4 +1,4 @@
 export * from './shared'
+export * from './default-theme'
 export * from '../dist/node/index'
 export * from '../dist/client/index'
-export * from '../dist/client/theme-default/config'
